fixup the logtracker app to scan the dict object for new extra keys and only log the ones that are new
this will ikely need to be fixed up to use a more targeted/specific logging app to get more meaningful columns for us to use for HQ purposes.

also added secondary threshold for actual alert messages.

set a dblog threshold for what to log to the db.
then an actual alert threshold, which should actually be higher priority or something to actually send the nags.